Add a `reason` field to `DisconnectedError`
It is useful in some applications to know why the connection was disconnected. This keeps track of the disconnection reason provided by the Softdevice and returns it in the `DisconnectedError`.

A new method `Connection::disconnect_with_reason()` has also been added to allow the user to indicate the reason for a locally initiated disconnection to the remote device.